# bank-employees-salaries-api/resources/openapi.yaml
openapi: 3.0.1
info:
  title: API для работы с данными посещаемости, отделами и сотрудниками
  version: 1.0.0
  description: |
    Система формирования отчетов для AES.
    Включает компоненты для работы с моделями и DTO.
servers:
  - url: http://localhost:8080/api
    description: Локальный сервер разработки
  - url: https://api.example.com/v1
    description: Продуктивный сервер
...
paths:
...
  /positions:
    post:
      operationId: addPosition
      tags:
        - Positions
      summary: Добавление новой должности
      description: Создает новую должность в системе
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'PositionDto.yaml#/components/schemas/Position'
      responses:
        '200':
          description: Успешное создание должности
          content:
            application/json:
              schema:
                $ref: 'PositionDto.yaml#/components/schemas/PositionDto'
        '400':
          description: Неверные входные данные
        '500':
          description: Внутренняя ошибка сервера

    get:
      operationId: getPositions
      tags:
        - Positions
      summary: Получение списка всех должностей
      description: Возвращает список всех должностей в системе
      responses:
        '200':
          description: Успешное получение списка должностей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: 'PositionDto.yaml#/components/schemas/PositionDto'
        '500':
          description: Внутренняя ошибка сервера

    delete:
      operationId: deleteAllPositions
      tags:
        - Positions
      summary: Удаление всех должностей
      description: Удаляет все должности из системы
      responses:
        '204':
          description: Все должности успешно удалены
        '500':
          description: Внутренняя ошибка сервера
  /positions/{id}:
    patch:
      operationId: updatePosition
      tags:
        - Positions
      summary: Обновление должности
      description: Обновляет информацию о должности по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'PositionDto.yaml#/components/schemas/Position'
      responses:
        '200':
          description: Успешное обновление должности
          content:
            application/json:
              schema:
                $ref: 'PositionDto.yaml#/components/schemas/PositionDto'
        '400':
          description: Неверные входные данные
        '404':
          description: Должность не найдена
        '500':
          description: Внутренняя ошибка сервера

    delete:
      operationId: deletePositionById
      tags:
        - Positions
      summary: Удаление должности по ID
      description: Удаляет должность по указанному ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        '204':
          description: Должность успешно удалена
        '404':
          description: Должность не найдена
        '500':
          description: Внутренняя ошибка сервера

    get:
      operationId: getPositionById
      tags:
        - Positions
      summary: Получение должности по ID
      description: Возвращает информацию о должности по указанному ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        '200':
          description: Успешное получение должности
          content:
            application/json:
              schema:
                $ref: 'PositionDto.yaml#/components/schemas/PositionDto'
        '404':
          description: Должность не найдена
        '500':
          description: Внутренняя ошибка сервера

  # ========== Salary Endpoints ==========
  /salaries:
    post:
      operationId: addSalary
      tags:
        - Salaries
      summary: Сохранить данные о зарплате
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'SalaryDto.yaml#/components/schemas/Salary'
      responses:
        '200':
          description: Данные о зарплате успешно сохранены
          content:
            application/json:
              schema:
                $ref: 'SalaryDto.yaml#/components/schemas/SalaryDto'

    delete:
      operationId: deleteAllSalaries
      tags:
        - Salaries
      summary: Удаление всех зарплат
      responses:
        '204':
          description: Все данные о зарплатах успешно удалены
  /salaries/period:
    get:
      operationId: getCompanySalariesByPeriod
      tags:
        - Salaries
      summary: Получить данные о зарплатах по компании за месяц
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'CommonDto.yaml#/components/schemas/DateTimePeriod'
      responses:
        '200':
          description: Список зарплат по компании за указанный месяц
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: 'SalaryDto.yaml#/components/schemas/SalaryDto'
  /salaries/{id}:
    patch:
      operationId: updateSalary
      tags:
        - Salaries
      summary: Обновить данные о зарплате
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
          description: ID записи о зарплате
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'SalaryDto.yaml#/components/schemas/Salary'
      responses:
        '200':
          description: Данные о зарплате успешно обновлены
          content:
            application/json:
              schema:
                $ref: 'SalaryDto.yaml#/components/schemas/SalaryDto'

    delete:
      operationId: deleteSalaryById
      tags:
        - Salaries
      summary: Удалить данные о зарплате по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
          description: ID записи о зарплате
      responses:
        '204':
          description: Данные о зарплате успешно удалены

    get:
      operationId: getSalaryById
      tags:
        - Salaries
      summary: Получить данные о зарплате по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
          description: ID записи о зарплате
      responses:
        '200':
          description: Данные о зарплате успешно получены
          content:
            application/json:
              schema:
                $ref: 'SalaryDto.yaml#/components/schemas/SalaryDto'
  /salaries/calculate/{employeeId}/period:
    get:
      operationId: getEmployeeSalaryByPeriod
      tags:
        - Salaries
      summary: Получить зарплату сотрудника за период
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
          description: ID сотрудника
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'CommonDto.yaml#/components/schemas/DateTimePeriod'
      responses:
        '200':
          description: Данные о зарплате сотрудника за месяц успешно получены
          content:
            application/json:
              schema:
                $ref: 'SalaryDto.yaml#/components/schemas/SalaryDto'
  /salaries/calculate/{departmentId}/period:
    get:
      operationId: getDepartmentSalariesByPeriod
      tags:
        - Salaries
      summary: Получить зарплаты отдела за период
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
          description: ID отдела
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'CommonDto.yaml#/components/schemas/DateTimePeriod'
      responses:
        '200':
          description: Список зарплат отдела за указанный месяц успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: 'SalaryDto.yaml#/components/schemas/SalaryDto'

  # ========== Salaries Data Endpoints ==========
  /salaries-dates:
    post:
      tags:
        - SalariesData
      summary: Добавить данные о зарплате в БД
      operationId: addSalariesData
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'SalariesDataDto.yaml#/components/schemas/SalariesData'
      responses:
        '200':
          description: Данные о зарплате успешно добавлены
          content:
            application/json:
              schema:
                $ref: 'SalariesDataDto.yaml#/components/schemas/SalariesDataDto'

    delete:
      tags:
        - SalariesData
      summary: Удалить все данные зарплат из БД
      operationId: deleteAllSalariesData
      responses:
        '204':
          description: Все данные о зарплатах успешно удалены

    get:
      tags:
        - SalariesData
      summary: Получить все данные зарплат в БД
      operationId: getAllSalariesData
      responses:
        '200':
          description: Список всех данных о зарплатах
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: 'SalariesDataDto.yaml#/components/schemas/SalariesDataDto'
  /salaries-dates/{id}:
    patch:
      tags:
        - SalariesData
      summary: Заменить данные о зарплате в БД
      operationId: updateSalariesDataById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'SalariesDataDto.yaml#/components/schemas/SalariesData'
      responses:
        '200':
          description: Данные о зарплате успешно обновлены
          content:
            application/json:
              schema:
                $ref: 'SalariesDataDto.yaml#/components/schemas/SalariesDataDto'

    delete:
      tags:
        - SalariesData
      summary: Удалить данные зарплаты в БД по id
      operationId: deleteSalariesDataById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        '204':
          description: Данные о зарплате успешно удалены

    get:
      tags:
        - SalariesData
      summary: Получить данные зарплаты в БД по id
      operationId: getSalariesDataById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        '200':
          description: Данные о зарплате
          content:
            application/json:
              schema:
                $ref: 'SalariesDataDto.yaml#/components/schemas/SalariesDataDto'

  # ========== Work Schedule Endpoints ==========
...